Changed variable and method names to be more concise. Added in a check to skip win con checking if someone folded. Added in TODO statements where necessary. Refactored and explained the straight flush checking. Removed Return winnings method from dealer class... unnecessary.
Corrected additional comma after the last cpu name is listed in the game log.

Added in rank variable in player class to hold rank of individual players for easier reference. Changed method names in player class, added in getter/setter for rank, and refactored how winnings are added or subtracted from a player who bets.

Changed the player constructors to assign ranks on creation. Added in comments and TODO statements. Removed commented out code. Changed bet field to a 0 instead of "type here". Implement way to show flop, turn, and river cards, and the cpu player cards at the end of the game. Check for win condition, and give winnings to the winner.